requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
av==14.4.0
blosc2==3.3.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cors==1.0.1
distro==1.9.0
docstring_parser==0.16
eval_type_backport==0.2.2
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.2
fsspec==2025.5.1
future==1.0.0
gevent==25.5.1
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.27.0
google-genai==1.19.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
livekit==1.0.8
livekit-agents==1.0.23
livekit-api==1.0.2
livekit-plugins-google==1.0.23
livekit-plugins-groq==1.0.23
livekit-plugins-noise-cancellation==0.2.4
livekit-plugins-openai==1.0.23
livekit-plugins-silero==1.0.23
livekit-plugins-turn-detector==1.0.23
livekit-protocol==1.0.3
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
mss==10.0.0
multidict==6.4.4
multimodal==0.0.13
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==3.5
numexpr==2.10.2
numpy==2.3.0
onnxruntime==1.22.0
openai==1.84.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pySmartDL==1.3.4
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rsa==4.9.1
safetensors==0.5.3
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.46.2
sympy==1.14.0
tables==3.10.2
tldextract==5.3.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
types-protobuf==4.25.0.20240417
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
zope.event==5.0
zope.interface==7.2
